+++ examples/simple.rs
@@ -18,7 +18,7 @@
 
 use std::convert::TryInto;
 
-use astarte_sdk::{types::AstarteType, AstarteOptions};
+use astarte_sdk::{builder::AstarteBuilder, types::AstarteType};
 use structopt::StructOpt;
 
 #[derive(Debug, StructOpt)]
@@ -49,7 +49,7 @@ async fn main() {
     } = Cli::from_args();
 
     let mut sdk_options =
-        AstarteOptions::new(&realm, &device_id, &credentials_secret, &pairing_url);
+        AstarteBuilder::new(&realm, &device_id, &credentials_secret, &pairing_url);
 
     sdk_options
         .add_interface_files("./examples/interfaces")
