+++ src/interface.rs
@@ -373,11 +373,11 @@ mod tests {
             reliability: Reliability::Unreliable,
         };
 
-        assert_eq!(mapping.is_compatible("/foo/value"), true);
-        assert_eq!(mapping.is_compatible("/bar/value"), true);
-        assert_eq!(mapping.is_compatible("/value"), false);
-        assert_eq!(mapping.is_compatible("/foo/bar/value"), false);
-        assert_eq!(mapping.is_compatible("/foo/value/bar"), false);
+        assert!(mapping.is_compatible("/foo/value"));
+        assert!(mapping.is_compatible("/bar/value"));
+        assert!(!mapping.is_compatible("/value"));
+        assert!(!mapping.is_compatible("/foo/bar/value"));
+        assert!(!mapping.is_compatible("/foo/value/bar"));
         assert_eq!(mapping.endpoint(), "/%{sensor_id}/value");
         assert_eq!(mapping.mapping_type(), MappingType::Double);
         assert_eq!(mapping.description(), Some("Mapping description"));
